refactor tour and add tour for cancel rule
